Flow and action changes for 802.1AD
ToDo: The flow structure needs to be updated to hold both the inner and outer
tpid since these are encoded by the kernel module.

The flow structure is updated to hold the customer tci.
Flow key extraction is changed to add support for ctci and both TPIDs.
Parsing to support pushing and popping with both single and double
tagged vlans. In response to reviewers comments on V6, all changes
affected by the flow structure are gathered in this patch.

Signed-off-by: Thomas F Herbert <thomasfherbert@gmail.com>